+++ test/test_helper.rb
@@ -3,6 +3,7 @@ require "minitest/reporters"
 Minitest::Reporters.use!(Minitest::Reporters::SpecReporter.new)
 require_relative "../lib/parse/stack.rb"
 require "minitest/autorun"
+require "minitest/mock"
 
 Parse.use_shortnames!